api1/requirements.txt
cryptography==3.4.6
fastapi==0.63.0
httpx==0.17.1
pyJWT==2.0.1
uvicorn[standard]==0.13.4
